matrix-gateway/requirements.txt
fastapi==0.100.0
matrix-nio==0.20.0
python-dotenv==1.0.0
tenacity==9.0.0
uvicorn==0.22.0
httpx==0.23.0
httpcore==0.15.0
h11==0.12.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-cov==6.0.0
redis==5.2.0
coverage==7.6.8
pydantic==1.10.2